--- .github/workflows/bot-audit.yml
@@ -1,78 +1,0 @@
-name: bot-audit
-
-on:
-  workflow_dispatch: null
-
-jobs:
-  audit:
-    name: audit
-    runs-on: "ubuntu-latest"
-    concurrency: audit
-    defaults:
-      run:
-        shell: bash -l {0}
-
-    steps:
-      - uses: actions/checkout@v3
-        with:
-          path: cf-scripts
-
-      - uses: conda-incubator/setup-miniconda@v2
-        with:
-          channels: conda-forge,defaults
-          channel-priority: strict
-          show-channel-urls: true
-          miniforge-version: latest
-          miniforge-variant: Mambaforge
-          python-version: ${{ vars.PYVER }}
-          use-mamba: true
-          environment-file: cf-scripts/environment.yml
-          activate-environment: cf-scripts
-
-      - name: do local setup and stop me if needed
-        run: |
-          pushd cf-scripts
-          python autotick-bot/stop_me_if_needed.py
-          popd
-
-      - name: install bot code
-        if: success() && ! env.CI_SKIP
-        run: |
-          source cf-scripts/autotick-bot/install_bot_code.sh
-        env:
-          PASSWORD: ${{ secrets.AUTOTICK_BOT_TOKEN }}
-
-      - name: audit dependencies
-        if: success() && ! env.CI_SKIP
-        run: |
-          pushd cf-graph
-
-          conda-forge-tick --run 5
-
-          popd
-        env:
-          PASSWORD: ${{ secrets.AUTOTICK_BOT_TOKEN }}
-
-      - name: deploy
-        if: github.ref == 'refs/heads/master' && ! cancelled() && ! env.CI_SKIP
-        run: |
-          pushd cf-graph
-
-          export CIRCLE_BUILD_URL="https://github.com/regro/cf-scripts/actions/runs/${RUN_ID}"
-          export CIRCLE_BUILD_NUM="actually-actions-${RUN_ID}"
-          conda-forge-tick --run -1
-
-          popd
-        env:
-          PASSWORD: ${{ secrets.AUTOTICK_BOT_TOKEN }}
-          RUN_ID: ${{ github.run_id }}
-
-      - name: bump on fail
-        if: github.ref == 'refs/heads/master' && failure() && ! env.CI_SKIP
-        run: |
-          export ACTION_URL="https://github.com/regro/cf-scripts/actions/runs/${RUN_ID}"
-          python cf-scripts/autotick-bot/bump_bot_team.py
-        env:
-          PASSWORD: ${{ secrets.AUTOTICK_BOT_TOKEN }}
-          RUN_ID: ${{ github.run_id }}
-          ACTION_NAME: ${{ github.workflow }}
